+++ packages/svelte-query/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.20](https://github.com/wundergraph/wundergraph/compare/@wundergraph/svelte-query@0.2.19...@wundergraph/svelte-query@0.2.20) (2023-05-04)
+
+**Note:** Version bump only for package @wundergraph/svelte-query
+
 ## [0.2.19](https://github.com/wundergraph/wundergraph/compare/@wundergraph/svelte-query@0.2.18...@wundergraph/svelte-query@0.2.19) (2023-05-03)
 
 **Note:** Version bump only for package @wundergraph/svelte-query